Fixes silent crashes when working with newer versions of pygame
Apparently newer versions of pygame need slightly different parameters.

Excerpt from ~/.slingshot/logfile.txt:
Traceback (most recent call last):
  File "slingshot.py", line 852, in <module>
    if __name__ == '__main__': main()
  File "slingshot.py", line 849, in main
    game = Game()
  File "slingshot.py", line 177, in __init__
    self.game_init()
  File "slingshot.py", line 181, in game_init
    self.round_init()
  File "slingshot.py", line 242, in round_init
    self.planetsprites = self.create_planets()
  File "slingshot.py", line 306, in create_planets
    result.add(Planet(result, self.background))
  File "/home/sb/OSS/slingshot/slingshot/planet.py", line 60, in __init__
    self.orig = pygame.transform.scale(self.image, (s, s))
TypeError: integer argument expected, got float

+++ slingshot/slingshot.py
@@ -441,7 +441,7 @@ class Game:
 				tmp.set_colorkey((0,0,0))
 				tmp = tmp.convert_alpha()
 				rect = tmp.get_rect()
-				s = (100 - self.show_round) * rect.h / 25
+				s = int((100 - self.show_round) * rect.h / 25)
 				tmp = pygame.transform.scale(tmp, (rect.w / rect.h * s, s ))
 				rect = tmp.get_rect()
 				rect.center = (399,299)
